Introduce serial_changeset.
serial_changeset allows callers to iterate over a serialized version
of a changeset (i.e., a set of column replacements). A serial_changeset
is about 40-50% faster to wrangle directly, as compared to the old story,
which parsed a serialized changeset into a change_t.

Support serial_changesets, and other similar changesets, via templates.