Internal: Add hooks for "does the parent really exist?".
We are using null pointers for "this node has no parent [within
its layer]". But we might want eventually to store backpointers from
one layer to the next higher layer. (Imagine an iterator.) So add some
hooks to hide the identification of null pointer with no-parent. These
hooks should be completely optimized away.