Msgpack-based RPC.
Requires some finagling particularly in server, because we now
read a big buffer from the kernel to userlevel, so we can't
simply call select() again (we must process the buffer to completion).

This is significantly faster than the old RPC. The reason is
basically system calls. Because we read in a big buffer server,
we can make fewer system calls, and we can flush less frequently on
tcp.