locked_tcursor: Store the root, not the basic_table.
This requires some rearrangement. So clean up table::root *outside*
the insert callback; instead clean up when initializing the tcursor.
The point of this is to avoid putting a pointer to the table in an RCU
callback. We don't want that, because perhaps the table has changed
(been destroy()ed).