Mark a node as inserting when changing a value to a new layer.
This is necessary to preserve an invariant useful for transactional
scan: if the collection of keys covered by a node changes, then that
node's full_version_value() must have changed too.

It also lets us simplify the protocol for installing the new layer.
Rather than introduce a new type of layer (unstable_layer), just
mark the node as inserting and everyone will retry.