mtd: add an option for configuring epoch interval
Current mtd increases epoch in fixed interval (1s) and doesn't provide
a way to configure it. However, the interval is an important parameter
for the epoch based reclamation of mastree. For example, the parameter
will affect a size of the limbo lists. Therefore it will affect space
efficiency and latency of masstree.

This commit adds a new option --epoch-interval to mtd for making
experiments with various epoch interval easy. Users can pass custom
epoch interval with the option in millisec unit.

Example usage (a case of 100ms interval):
$ ./mtd --epoch-interval 100