mtd: asynchronous rcu quiesce for reducing latency
This commit adds a new option --async-quiesce to mtd. This option
enables asynchronous rcu quiesce for reducing request latency. The
strategy is very simple: a thread collects garbage objects of its
limbo list (which can be simply free()ed, not returned into
threadinfo.pool_) in a vector and passes to a thread that only does
reclamation.