mtd: stop RCU quiesce state after serializing a result of a request
Current mtd calls ti->rcu_stop() before serializing a result of a
request. However, IIUC, the result, e.g. read value of get request, is
managed under RCU and is not deep copied for the thread. So calling
ti->rcu_stop() will give a chance of collecting a value that can be
contained by the result to other threads. It would result a situation
that a request processing thread can read stale pointers.

This commit swaps an order of ti->rcu_stop() and result serialization
for avoid the problem.