*: configurable RCU free count
The commit fd8c6782bb74 introduced a limit of reclaimed object in
single RCU quiesce. However, the feature degrades throughput
unconditionally as reported in
https://github.com/kohler/masstree-beta/pull/21#issuecomment-204379130

This commit adds a new option --rcu-free-count to mtd. The option is
used as the limit of reclamation. If the option isn't passed, the
limit doesn't exist (same to the mtd before fd8c6782bb74).