New remove plan.
There were bugs with the old remove plan, demonstrated by unit-mt.
This plan aims to combine the operation of collapse() and the
previous walk up the tree. It makes more sense to me.

In this plan, deleted internodes have all their child pointers set
to nullptr. This is important because reach_leaf() doesn't check
deleted bits in the version number; only nullptr tells reach_leaf()
to retry from the root.