Fix #40.
Or at least maybe do so.

Since a nullptr child in an internal node signals readers that they
must retry, such internodes must be preserved for at least an RCU
grace period. The old code could get rid of such internodes, because
it prematurely deleted internodes with 2 children if the 1st child
was nullptr.

This code does not delete such internodes. That's a bit of a shame
because such internodes are useless. An RCU callback could clean
them up; don't bother with one yet.